requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
attrs==19.3.0
black==19.10b0
click==7.1.2
gast==0.3.3
grpcio==1.30.0
h5py==2.10.0
importlib-metadata==1.7.0
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mock==4.0.2
numpy==1.19.0
opencv-python==4.3.0.38
pathspec==0.8.0
protobuf==3.12.2
regex==2020.7.14
six==1.15.0
tensorboard==1.11
tensorflow-gpu==1.11
tensorflow-estimator==1.11
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
zipp==3.1.0
